+++ delivery_app/lib/screens/merchant/merchant_pages/merchant_order_page.dart
@@ -1,4 +1,5 @@
 import 'package:delivery_app/customWidgets/FancyButton.dart';
+import 'package:delivery_app/customWidgets/merchantDrawer.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
@@ -8,7 +9,6 @@ class MerchantOrder extends StatefulWidget {
 }
 
 class _MerchantOrderState extends State<MerchantOrder> {
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -20,15 +20,7 @@ class _MerchantOrderState extends State<MerchantOrder> {
         ),
         centerTitle: true,
       ),
-      drawer: Drawer(
-        child: ListView(
-          children: [
-            DrawerHeader(child: Text('Hello')),
-            ListTile(title: Text('A')),
-            ListTile(title: Text('B')),
-          ],
-        ),
-      ),
+      drawer: MerchantDrawer(),
       backgroundColor: Colors.white,
       body: Column(
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
@@ -38,12 +30,12 @@ class _MerchantOrderState extends State<MerchantOrder> {
           SizedBox(),
           Center(
               child: Container(
-                margin: EdgeInsets.fromLTRB(0, 250, 0, 250),
-                child: Text(
-                  'No orders yet',
-                  style: TextStyle(fontSize: 40.0, color: Colors.orange[500]),
-                ),
-              )),
+            margin: EdgeInsets.fromLTRB(0, 250, 0, 250),
+            child: Text(
+              'No orders yet',
+              style: TextStyle(fontSize: 40.0, color: Colors.orange[500]),
+            ),
+          )),
           Container(
             margin: EdgeInsets.fromLTRB(20, 0, 20, 0),
             child: FancyButton(
