+++ delivery_app/lib/screens/merchant/merchant_screens/merchant_registers_driver_screen.dart
@@ -1,33 +1,32 @@
-import 'package:delivery_app/merchant_home_page.dart';
 import 'package:flutter/material.dart';
-import 'package:delivery_app/models/merchant_model.dart';
-import 'package:delivery_app/url_link.dart';
-import 'dart:convert';
+import 'package:delivery_app/models/driver_model.dart';
 import 'package:http/http.dart' as http;
+import 'dart:convert';
+import 'package:delivery_app/url_link.dart';
 
-// Merchant: full name, phone number, company name, company address.
-class MerchantSignUp extends StatefulWidget {
+// Driver: full name, phone number, address.
+class MerchantRegisterDriver extends StatefulWidget {
+  final String merchantID;
+  MerchantRegisterDriver({Key key, @required this.merchantID})
+      : super(key: key);
   @override
-  _MerchantSignUpState createState() => _MerchantSignUpState();
+  _MerchantRegisterDriverState createState() => _MerchantRegisterDriverState();
 }
 
-class _MerchantSignUpState extends State<MerchantSignUp> {
-  Future<MerchantModel> addMerchant(String firstName, String lastName,
-      String companyName, String companyAddress, String phoneNumber) async {
-    //http://a84a794b3db6.ngrok.io/
+class _MerchantRegisterDriverState extends State<MerchantRegisterDriver> {
+  Future<DriverModel> addDriver(
+      String firstName, String lastName, String phoneNumber) async {
     var uri = Uri(
       scheme: 'https',
-      host: theLink,
-      path: '/register/driver/39',
+      host: ngrokLink,
+      path: '/driver/register/${widget.merchantID}',
     );
-    assert(
-        uri.toString() == 'https://$theLink/register/merchant');
+    assert(uri.toString() ==
+        'https://$ngrokLink/driver/register/${widget.merchantID}');
 
     Map<String, String> a = {
       "first_name": firstName,
       "last_name": lastName,
-      "company_name": companyName,
-      "company_address": companyAddress,
       "phone_number": phoneNumber
     };
     var b = json.encode(a);
@@ -35,15 +34,13 @@ class _MerchantSignUpState extends State<MerchantSignUp> {
     http.Response response = await http
         .post(uri, body: b, headers: {"content-type": "application/json"});
     final String responseString = response.body;
-    return merchantModelFromJson(responseString);
+    return driverModelFromJson(responseString);
   }
 
-  MerchantModel _merchant;
+  DriverModel _driver;
 
   final TextEditingController firstNameController = TextEditingController();
   final TextEditingController lastNameController = TextEditingController();
-  final TextEditingController companyNameController = TextEditingController();
-  final TextEditingController companyAddressController = TextEditingController();
   final TextEditingController phoneNumberController = TextEditingController();
 
   @override
@@ -70,16 +67,12 @@ class _MerchantSignUpState extends State<MerchantSignUp> {
               onPressed: () async {
                 final String firstName = firstNameController.text;
                 final String lastName = lastNameController.text;
-                final String companyName = companyNameController.text;
-                final String companyAddress = companyAddressController.text;
                 final String phoneNumber = phoneNumberController.text;
-                final MerchantModel newMerchant = await addMerchant(firstName,
-                    lastName, companyName, companyAddress, phoneNumber);
+                final DriverModel newDriver =
+                    await addDriver(firstName, lastName, phoneNumber);
                 setState(() {
-                  _merchant = newMerchant;
+                  _driver = newDriver;
                 });
-
-                Navigator.of(context).pushNamed('/merchantPage');
               },
               child: Text(
                 "Done",
@@ -93,7 +86,7 @@ class _MerchantSignUpState extends State<MerchantSignUp> {
         ],
         automaticallyImplyLeading: false,
         title: Text(
-          "Sign up",
+          "Add Driver",
           style: TextStyle(color: Colors.white, fontSize: 23),
         ),
         centerTitle: true,
@@ -152,41 +145,23 @@ class _MerchantSignUpState extends State<MerchantSignUp> {
                         ),
                       ),
                     ),
-                    Container(
-                      width: screenSize * 0.85,
-                      child: TextField(
-                        controller: companyNameController,
-                        style: TextStyle(fontSize: 22),
-                        decoration: new InputDecoration(
-                          fillColor: Color(0xFFF8F8F8),
-                          filled: true,
-                          contentPadding: EdgeInsets.only(left: 15.0),
-                          hintText: 'Company name',
-                          border: new OutlineInputBorder(
-                            borderRadius: const BorderRadius.all(
-                              const Radius.circular(10.0),
-                            ),
-                          ),
-                        ),
-                      ),
-                    ),
-                    Container(
-                        width: screenSize * 0.85,
-                        child: TextField(
-                          controller: companyAddressController,
-                          style: TextStyle(fontSize: 22),
-                          decoration: new InputDecoration(
-                            fillColor: Color(0xFFF8F8F8),
-                            filled: true,
-                            contentPadding: EdgeInsets.only(left: 15.0),
-                            hintText: 'Company address',
-                            border: new OutlineInputBorder(
-                              borderRadius: const BorderRadius.all(
-                                const Radius.circular(10.0),
-                              ),
-                            ),
-                          ),
-                        )),
+                    // Container(
+                    //   width: screenSize * 0.85,
+                    //   child: TextField(
+                    //     style: TextStyle(fontSize: 22),
+                    //     decoration: new InputDecoration(
+                    //       fillColor: Color(0xFFF8F8F8),
+                    //       filled: true,
+                    //       contentPadding: EdgeInsets.only(left: 15.0),
+                    //       hintText: 'Home address',
+                    //       border: new OutlineInputBorder(
+                    //         borderRadius: const BorderRadius.all(
+                    //           const Radius.circular(10.0),
+                    //         ),
+                    //       ),
+                    //     ),
+                    //   ),
+                    // ),
                     Container(
                         width: screenSize * 0.85,
                         child: TextField(
@@ -208,14 +183,8 @@ class _MerchantSignUpState extends State<MerchantSignUp> {
                 )),
             Container(
               margin: EdgeInsets.only(
-                top: screenSize * 0.35,
+                top: screenSize * 0.5,
               ),
-              child: FlatButton(
-                  onPressed: () => {Navigator.of(context).pushNamed('/merchantSignInScreen')},
-                  child: Text(
-                    "Already have an account? Sign in",
-                    style: TextStyle(fontSize: 20, color: Color(0xFF8D8D8D)),
-                  )),
             )
           ]))),
       backgroundColor: Color(0xFFF6F8F5),
